[05.11.2024] Moved crash detection to separate file, merge with main
