+++ rescue_now_app/test/widget_test.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import 'package:rescue_now_app/main.dart'; // Update this to the correct path if necessary
+import 'package:rescue_now_app/crash_detection.dart'; // Update this to the correct path if necessary
 
 void main() {
   testWidgets('Crash Detection App Test', (WidgetTester tester) async {
